:building_construction: Refactored base project for heroku migration
