+++ api/users/user-model.js
@@ -1,4 +1,4 @@
-const db = require('../../data/dbConfig.js');
+const db = require('../../data/db-config.js');
 
 module.exports = {
   getUsers,
