+++ data/seeds/002-users.js
@@ -1,3 +1,5 @@
+const bcrypt = require('bcryptjs');
+
 exports.seed = function(knex) {
   return knex('users')
     .del()
