+++ package.json
@@ -1,30 +1,31 @@
 {
-  "name": "backend",
+  "name": "webapi-blog",
   "version": "1.0.0",
-  "description": "A backend for music meteorologist powered by node.js, express.js using a postgresql database for persistance.",
+  "description": "Web API Blog Project",
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "development": "nodemon index.js",
+    "server": "nodemon index.js",
     "test": "cross-env DB_ENV=development jest --watch"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/labs15-music-meteorologist/backend.git"
+    "url": "git+https://github.com/LambdaSchool/Node-Blog.git"
   },
   "keywords": [
     "node",
     "express",
-    "spotify",
-    "postgresql"
+    "blog",
+    "api"
   ],
-  "author": "Justin Chen, Ben Feole, Sascha Majewsky",
-  "license": "MIT",
+  "author": "Lambda School",
+  "license": "ISC",
   "bugs": {
-    "url": "https://github.com/labs15-music-meteorologist/backend/issues"
+    "url": "https://github.com/LambdaSchool/Node-Blog/issues"
   },
-  "homepage": "https://github.com/labs15-music-meteorologist/backend#readme",
+  "homepage": "https://github.com/LambdaSchool/Node-Blog#readme",
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "cross-env": "^5.2.0",
     "dotenv": "^8.0.0",
@@ -34,12 +35,9 @@
     "knex": "^0.18.4",
     "knex-cleaner": "^1.2.1",
     "moment": "^2.24.0",
-    "morgan": "^1.9.1",
-    "oauth": "^0.9.15",
-    "passport": "^0.4.0",
-    "passport-spotify": "^1.1.0",
-    "pg": "^7.12.1",
-    "postgresql": "0.0.1"
+    "pg": "^7.12.0",
+    "postgresql": "0.0.1",
+    "sqlite3": "^4.0.9"
   },
   "devDependencies": {
     "jest": "^24.8.0",
