--- src/data/dbConfig.js
@@ -1,8 +1,0 @@
-const knex = require('knex');
-/* require('dotenv').config(); */
-
-const knexConfig = require('./../../knexfile');
-
-const environment = process.env.DB_ENV;
-
-module.exports = knex(knexConfig[environment]);
